Enum gdk_pixbuf::InterpType

pub enum InterpType {
    Nearest,
    Tiles,
    Bilinear,
    Hyper,
}

This enumeration describes the different interpolation modes that can be used with the scaling functions. InterpType::Nearest is the fastest scaling method, but has horrible quality when scaling down. InterpType::Bilinear is the best choice if you aren't sure what to choose, it has a good speed/quality balance.

<note> Cubic filtering is missing from the list; hyperbolic interpolation is just as fast and results in higher quality. </note>

Variants

Nearest

Nearest neighbor sampling; this is the fastest and lowest quality mode. Quality is normally unacceptable when scaling down, but may be OK when scaling up.

Tiles

This is an accurate simulation of the PostScript image operator without any interpolation enabled. Each pixel is rendered as a tiny parallelogram of solid color, the edges of which are implemented with antialiasing. It resembles nearest neighbor for enlargement, and bilinear for reduction.

Bilinear

Best quality/speed balance; use this mode by default. Bilinear interpolation. For enlargement, it is equivalent to point-sampling the ideal bilinear-interpolated image. For reduction, it is equivalent to laying down small tiles and integrating over the coverage area.

Hyper

This is the slowest and highest quality reconstruction function. It is derived from the hyperbolic filters in Wolberg's "Digital Image Warping", and is formally defined as the hyperbolic-filter sampling the ideal hyperbolic-filter interpolated image (the filter is designed to be idempotent for 1:1 pixel mapping).