Trait gdk_pixbuf::prelude::PixbufSimpleAnimExt

pub trait PixbufSimpleAnimExt {
    fn add_frame(&self, pixbuf: &Pixbuf);

    fn get_loop(&self) -> bool;

    fn set_loop(&self, loop_: bool);

    fn connect_property_loop_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all PixbufSimpleAnim methods.

Implementors

PixbufSimpleAnim

Required Methods

fn add_frame(&self, pixbuf: &Pixbuf)

Adds a new frame to self. The pixbuf must have the dimensions specified when the animation was constructed.

pixbuf

the pixbuf to add

fn get_loop(&self) -> bool

Gets whether self should loop indefinitely when it reaches the end.

Returns

true if the animation loops forever, false otherwise

fn set_loop(&self, loop_: bool)

Sets whether self should loop indefinitely when it reaches the end.

loop_

whether to loop the animation

fn connect_property_loop_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId