Struct gdk_pixbuf::Pixbuf

pub struct Pixbuf(_, _);

This is the main structure in the gdk-pixbuf library. It is used to represent images. It contains information about the image's pixel data, its color space, bits per sample, width and height, and the rowstride (the number of bytes between the start of one row and the start of the next).

Implements

PixbufExt

Methods

impl Pixbuf

pub fn new(
colorspace: Colorspace,
has_alpha: bool,
bits_per_sample: i32,
width: i32,
height: i32
) -> Pixbuf

Creates a new Pixbuf structure and allocates a buffer for it. The buffer has an optimal rowstride. Note that the buffer is not cleared; you will have to fill it completely yourself.

colorspace

Color space for image

has_alpha

Whether the image should have transparency information

bits_per_sample

Number of bits per color sample

width

Width of image in pixels, must be > 0

height

Height of image in pixels, must be > 0

Returns

A newly-created Pixbuf with a reference count of 1, or None if not enough memory could be allocated for the image buffer.

pub fn new_from_bytes(
data: &Bytes,
colorspace: Colorspace,
has_alpha: bool,
bits_per_sample: i32,
width: i32,
height: i32,
rowstride: i32
) -> Pixbuf

Creates a new Pixbuf out of in-memory readonly image data. Currently only RGB images with 8 bits per sample are supported. This is the glib::Bytes variant of Pixbuf::new_from_data.

Feature: v2_32

data

Image data in 8-bit/sample packed format inside a glib::Bytes

colorspace

Colorspace for the image data

has_alpha

Whether the data has an opacity channel

bits_per_sample

Number of bits per sample

width

Width of the image in pixels, must be > 0

height

Height of the image in pixels, must be > 0

rowstride

Distance in bytes between row starts

Returns

A newly-created Pixbuf structure with a reference count of 1.

pub fn new_from_inline(data: &[u8], copy_pixels: bool) -> Result<Pixbuf, Error>

Create a Pixbuf from a flat representation that is suitable for storing as inline data in a program. This is useful if you want to ship a program with images, but don't want to depend on any external files.

gdk-pixbuf ships with a program called [gdk-pixbuf-csource][gdk-pixbuf-csource], which allows for conversion of GdkPixbufs into such a inline representation. In almost all cases, you should pass the --raw option to gdk-pixbuf-csource. A sample invocation would be:

 gdk-pixbuf-csource --raw --name=myimage_inline myimage.png

For the typical case where the inline pixbuf is read-only static data, you don't need to copy the pixel data unless you intend to write to it, so you can pass false for copy_pixels. (If you pass --rle to gdk-pixbuf-csource, a copy will be made even if copy_pixels is false, so using this option is generally a bad idea.)

If you create a pixbuf from const inline data compiled into your program, it's probably safe to ignore errors and disable length checks, since things will always succeed:

pixbuf = gdk_pixbuf_new_from_inline (-1, myimage_inline, FALSE, NULL);

For non-const inline data, you could get out of memory. For untrusted inline data located at runtime, you could have corrupt inline data in addition.

Deprecated since 2.32

Use gio::Resource instead.

data_length

Length in bytes of the data argument or -1 to disable length checks

data

Byte data containing a serialized Pixdata structure

copy_pixels

Whether to copy the pixel data, or use direct pointers data for the resulting pixbuf

Returns

A newly-created Pixbuf structure with a reference, count of 1, or None if an error occurred.

pub fn new_from_resource(resource_path: &str) -> Result<Pixbuf, Error>

Creates a new pixbuf by loading an image from an resource.

The file format is detected automatically. If None is returned, then error will be set.

resource_path

the path of the resource file

Returns

A newly-created pixbuf, or None if any of several error conditions occurred: the file could not be opened, the image format is not supported, there was not enough memory to allocate the image buffer, the stream contained invalid data, or the operation was cancelled.

pub fn new_from_resource_at_scale(
resource_path: &str,
width: i32,
height: i32,
preserve_aspect_ratio: bool
) -> Result<Pixbuf, Error>

Creates a new pixbuf by loading an image from an resource.

The file format is detected automatically. If None is returned, then error will be set.

The image will be scaled to fit in the requested size, optionally preserving the image's aspect ratio. When preserving the aspect ratio, a width of -1 will cause the image to be scaled to the exact given height, and a height of -1 will cause the image to be scaled to the exact given width. When not preserving aspect ratio, a width or height of -1 means to not scale the image at all in that dimension.

The stream is not closed.

resource_path

the path of the resource file

width

The width the image should have or -1 to not constrain the width

height

The height the image should have or -1 to not constrain the height

preserve_aspect_ratio

true to preserve the image's aspect ratio

Returns

A newly-created pixbuf, or None if any of several error conditions occurred: the file could not be opened, the image format is not supported, there was not enough memory to allocate the image buffer, the stream contained invalid data, or the operation was cancelled.

pub fn new_from_stream<'a, P: IsA<InputStream>, Q: Into<Option<&'a Cancellable>>>(
stream: &P,
cancellable: Q
) -> Result<Pixbuf, Error>

Creates a new pixbuf by loading an image from an input stream.

The file format is detected automatically. If None is returned, then error will be set. The cancellable can be used to abort the operation from another thread. If the operation was cancelled, the error gio::IOErrorEnum::Cancelled will be returned. Other possible errors are in the GDK_PIXBUF_ERROR and G_IO_ERROR domains.

The stream is not closed.

stream

a gio::InputStream to load the pixbuf from

cancellable

optional gio::Cancellable object, None to ignore

Returns

A newly-created pixbuf, or None if any of several error conditions occurred: the file could not be opened, the image format is not supported, there was not enough memory to allocate the image buffer, the stream contained invalid data, or the operation was cancelled.

pub fn new_from_stream_at_scale<'a, P: IsA<InputStream>, Q: Into<Option<&'a Cancellable>>>(
stream: &P,
width: i32,
height: i32,
preserve_aspect_ratio: bool,
cancellable: Q
) -> Result<Pixbuf, Error>

Creates a new pixbuf by loading an image from an input stream.

The file format is detected automatically. If None is returned, then error will be set. The cancellable can be used to abort the operation from another thread. If the operation was cancelled, the error gio::IOErrorEnum::Cancelled will be returned. Other possible errors are in the GDK_PIXBUF_ERROR and G_IO_ERROR domains.

The image will be scaled to fit in the requested size, optionally preserving the image's aspect ratio.

When preserving the aspect ratio, a width of -1 will cause the image to be scaled to the exact given height, and a height of -1 will cause the image to be scaled to the exact given width. If both width and height are given, this function will behave as if the smaller of the two values is passed as -1.

When not preserving aspect ratio, a width or height of -1 means to not scale the image at all in that dimension.

The stream is not closed.

stream

a gio::InputStream to load the pixbuf from

width

The width the image should have or -1 to not constrain the width

height

The height the image should have or -1 to not constrain the height

preserve_aspect_ratio

true to preserve the image's aspect ratio

cancellable

optional gio::Cancellable object, None to ignore

Returns

A newly-created pixbuf, or None if any of several error conditions occurred: the file could not be opened, the image format is not supported, there was not enough memory to allocate the image buffer, the stream contained invalid data, or the operation was cancelled.

pub fn new_from_xpm_data(data: &[&str]) -> Pixbuf

Creates a new pixbuf by parsing XPM data in memory. This data is commonly the result of including an XPM file into a program's C source.

data

Pointer to inline XPM data.

Returns

A newly-created pixbuf with a reference count of 1.

pub fn calculate_rowstride(
colorspace: Colorspace,
has_alpha: bool,
bits_per_sample: i32,
width: i32,
height: i32
) -> i32

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn get_formats() -> Vec<PixbufFormat>

Obtains the available information about the image formats supported by Pixbuf.

Returns

A list of GdkPixbufFormats describing the supported image formats. The list should be freed when it is no longer needed, but the structures themselves are owned by Pixbuf and should not be freed.

impl Pixbuf

pub fn new_from_vec(
vec: Vec<u8>,
colorspace: Colorspace,
has_alpha: bool,
bits_per_sample: i32,
width: i32,
height: i32,
row_stride: i32
) -> Pixbuf

pub fn new_from_file<T: AsRef<Path>>(filename: T) -> Result<Pixbuf, Error>

Creates a new pixbuf by loading an image from a file. The file format is detected automatically. If None is returned, then error will be set. Possible errors are in the GDK_PIXBUF_ERROR and G_FILE_ERROR domains.

filename

Name of file to load, in the GLib file name encoding

Returns

A newly-created pixbuf with a reference count of 1, or None if any of several error conditions occurred: the file could not be opened, there was no loader for the file's format, there was not enough memory to allocate the image buffer, or the image file contained invalid data.

pub fn new_from_file_at_size<T: AsRef<Path>>(
filename: T,
width: i32,
height: i32
) -> Result<Pixbuf, Error>

Creates a new pixbuf by loading an image from a file. The file format is detected automatically. If None is returned, then error will be set. Possible errors are in the GDK_PIXBUF_ERROR and G_FILE_ERROR domains.

The image will be scaled to fit in the requested size, preserving the image's aspect ratio. Note that the returned pixbuf may be smaller than width x height, if the aspect ratio requires it. To load and image at the requested size, regardless of aspect ratio, use Pixbuf::new_from_file_at_scale.

filename

Name of file to load, in the GLib file name encoding

width

The width the image should have or -1 to not constrain the width

height

The height the image should have or -1 to not constrain the height

Returns

A newly-created pixbuf with a reference count of 1, or None if any of several error conditions occurred: the file could not be opened, there was no loader for the file's format, there was not enough memory to allocate the image buffer, or the image file contained invalid data.

pub fn new_from_file_at_scale<T: AsRef<Path>>(
filename: T,
width: i32,
height: i32,
preserve_aspect_ratio: bool
) -> Result<Pixbuf, Error>

Creates a new pixbuf by loading an image from a file. The file format is detected automatically. If None is returned, then error will be set. Possible errors are in the GDK_PIXBUF_ERROR and G_FILE_ERROR domains. The image will be scaled to fit in the requested size, optionally preserving the image's aspect ratio.

When preserving the aspect ratio, a width of -1 will cause the image to be scaled to the exact given height, and a height of -1 will cause the image to be scaled to the exact given width. When not preserving aspect ratio, a width or height of -1 means to not scale the image at all in that dimension. Negative values for width and height are allowed since 2.8.

filename

Name of file to load, in the GLib file name encoding

width

The width the image should have or -1 to not constrain the width

height

The height the image should have or -1 to not constrain the height

preserve_aspect_ratio

true to preserve the image's aspect ratio

Returns

A newly-created pixbuf with a reference count of 1, or None if any of several error conditions occurred: the file could not be opened, there was no loader for the file's format, there was not enough memory to allocate the image buffer, or the image file contained invalid data.

pub fn new_from_stream_async<'a, P: IsA<InputStream>, Q: Into<Option<&'a Cancellable>>, R: FnOnce(Result<Pixbuf, Error>) + Send + 'static>(
stream: &P,
cancellable: Q,
callback: R
)

Creates a new pixbuf by asynchronously loading an image from an input stream.

For more details see Pixbuf::new_from_stream, which is the synchronous version of this function.

When the operation is finished, callback will be called in the main thread. You can then call Pixbuf::new_from_stream_finish to get the result of the operation.

stream

a gio::InputStream from which to load the pixbuf

cancellable

optional gio::Cancellable object, None to ignore

callback

a GAsyncReadyCallback to call when the pixbuf is loaded

user_data

the data to pass to the callback function

pub fn new_from_stream_at_scale_async<'a, P: IsA<InputStream>, Q: Into<Option<&'a Cancellable>>, R: FnOnce(Result<Pixbuf, Error>) + Send + 'static>(
stream: &P,
width: i32,
height: i32,
preserve_aspect_ratio: bool,
cancellable: Q,
callback: R
)

Creates a new pixbuf by asynchronously loading an image from an input stream.

For more details see Pixbuf::new_from_stream_at_scale, which is the synchronous version of this function.

When the operation is finished, callback will be called in the main thread. You can then call Pixbuf::new_from_stream_finish to get the result of the operation.

stream

a gio::InputStream from which to load the pixbuf

width

the width the image should have or -1 to not constrain the width

height

the height the image should have or -1 to not constrain the height

preserve_aspect_ratio

true to preserve the image's aspect ratio

cancellable

optional gio::Cancellable object, None to ignore

callback

a GAsyncReadyCallback to call when the pixbuf is loaded

user_data

the data to pass to the callback function

Important traits for &'a mut [u8]
impl<'a> Write for &'a mut [u8]impl<'a> Read for &'a [u8]

pub unsafe fn get_pixels(&self) -> &mut [u8]

pub fn put_pixel(&self, x: i32, y: i32, red: u8, green: u8, blue: u8, alpha: u8)

pub fn get_file_info<T: AsRef<Path>>(
filename: T
) -> Option<(PixbufFormat, i32, i32)>

Parses an image file far enough to determine its format and size.

filename

The name of the file to identify.

width

Return location for the width of the image, or None

height

Return location for the height of the image, or None

Returns

A PixbufFormat describing the image format of the file or None if the image format wasn't recognized. The return value is owned by Pixbuf and should not be freed.

pub fn get_file_info_async<'a, P: Into<Option<&'a Cancellable>>, Q: FnOnce(Result<Option<(PixbufFormat, i32, i32)>, Error>) + Send + 'static, T: AsRef<Path>>(
filename: T,
cancellable: P,
callback: Q
)

Asynchronously parses an image file far enough to determine its format and size.

For more details see Pixbuf::get_file_info, which is the synchronous version of this function.

When the operation is finished, callback will be called in the main thread. You can then call Pixbuf::get_file_info_finish to get the result of the operation.

Feature: v2_32

filename

The name of the file to identify

cancellable

optional gio::Cancellable object, None to ignore

callback

a GAsyncReadyCallback to call when the file info is available

user_data

the data to pass to the callback function

pub fn save_to_bufferv(
&self,
type_: &str,
options: &[(&str, &str)]
) -> Result<Vec<u8>, Error>

pub fn save_to_streamv<'a, P: IsA<OutputStream>, Q: Into<Option<&'a Cancellable>>>(
&self,
stream: &P,
type_: &str,
options: &[(&str, &str)],
cancellable: Q
) -> Result<(), Error>

pub fn savev<T: AsRef<Path>>(
&self,
filename: T,
type_: &str,
options: &[(&str, &str)]
) -> Result<(), Error>