Struct gdk_pixbuf::PixbufAnimation

pub struct PixbufAnimation(_, _);

An opaque struct representing an animation.

Implements

PixbufAnimationExt

Methods

impl PixbufAnimation

fn new_from_file(file: &str) -> Result<PixbufAnimation, Error>

Creates a new animation by loading it from a file. The file format is detected automatically. If the file's format does not support multi-frame images, then an animation with a single frame will be created. Possible errors are in the GDK_PIXBUF_ERROR and G_FILE_ERROR domains.

filename

Name of file to load, in the GLib file name encoding

Returns

A newly-created animation with a reference count of 1, or None if any of several error conditions ocurred: the file could not be opened, there was no loader for the file's format, there was not enough memory to allocate the image buffer, or the image file contained invalid data.

fn new_from_resource(resource_path: &str) -> Result<PixbufAnimation, Error>

Creates a new pixbuf animation by loading an image from an resource.

The file format is detected automatically. If None is returned, then error will be set.

Since: 2.28

resource_path

the path of the resource file

Returns

A newly-created animation, or None if any of several error conditions occurred: the file could not be opened, the image format is not supported, there was not enough memory to allocate the image buffer, the stream contained invalid data, or the operation was cancelled.