Struct gdk_pixbuf::PixbufAnimationIter

pub struct PixbufAnimationIter(_, _);

An opaque struct representing an iterator which points to a certain position in an animation.

Methods

impl PixbufAnimationIter

fn advance(&self, start_time: &TimeVal) -> bool

Possibly advances an animation to a new frame. Chooses the frame based on the start time passed to PixbufAnimationExt::get_iter.

current_time would normally come from g_get_current_time, and must be greater than or equal to the time passed to PixbufAnimationExt::get_iter, and must increase or remain unchanged each time PixbufAnimationIter::get_pixbuf is called. That is, you can't go backward in time; animations only play forward.

As a shortcut, pass None for the current time and g_get_current_time will be invoked on your behalf. So you only need to explicitly pass current_time if you're doing something odd like playing the animation at double speed.

If this function returns false, there's no need to update the animation display, assuming the display had been rendered prior to advancing; if true, you need to call PixbufAnimationIter::get_pixbuf and update the display with the new pixbuf.

current_time

current time

Returns

true if the image may need updating

fn get_pixbuf(&self) -> Pixbuf

Gets the current pixbuf which should be displayed; the pixbuf might not be the same size as the animation itself (PixbufAnimationExt::get_width, PixbufAnimationExt::get_height). This pixbuf should be displayed for PixbufAnimationIter::get_delay_time milliseconds. The caller of this function does not own a reference to the returned pixbuf; the returned pixbuf will become invalid when the iterator advances to the next frame, which may happen anytime you call PixbufAnimationIter::advance. Copy the pixbuf to keep it (don't just add a reference), as it may get recycled as you advance the iterator.

Returns

the pixbuf to be displayed

fn get_delay_time(&self) -> i32

Gets the number of milliseconds the current pixbuf should be displayed, or -1 if the current pixbuf should be displayed forever. g_timeout_add conveniently takes a timeout in milliseconds, so you can use a timeout to schedule the next update.

Note that some formats, like GIF, might clamp the timeout values in the image file to avoid updates that are just too quick. The minimum timeout for GIF images is currently 20 milliseconds.

Returns

delay time in milliseconds (thousandths of a second)

fn on_currently_loading_frame(&self) -> bool

Used to determine how to respond to the area_updated signal on PixbufLoader when loading an animation. area_updated is emitted for an area of the frame currently streaming in to the loader. So if you're on the currently loading frame, you need to redraw the screen for the updated area.

Returns

true if the frame we're on is partially loaded, or the last frame