Struct gdk_pixbuf::PixbufFormat

pub struct PixbufFormat(_);

Methods

impl PixbufFormat

pub fn get_description(&self) -> Option<String>

Returns a description of the format.

Returns

a description of the format.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn get_extensions(&self) -> Vec<String>

Returns the filename extensions typically used for files in the given format.

Returns

a None-terminated array of filename extensions which must be freed with g_strfreev when it is no longer needed.

pub fn get_license(&self) -> Option<String>

Returns information about the license of the image loader for the format. The returned string should be a shorthand for a wellknown license, e.g. "LGPL", "GPL", "QPL", "GPL/QPL", or "other" to indicate some other license. This string should be freed with g_free when it's no longer needed.

Returns

a string describing the license of self.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn get_mime_types(&self) -> Vec<String>

Returns the mime types supported by the format.

Returns

a None-terminated array of mime types which must be freed with g_strfreev when it is no longer needed.

pub fn get_name(&self) -> Option<String>

Returns the name of the format.

Returns

the name of the format.

pub fn is_disabled(&self) -> bool

Returns whether this image format is disabled. See PixbufFormat::set_disabled.

Returns

whether this image format is disabled.

pub fn is_save_option_supported(&self, option_key: &str) -> bool

Returns true if the save option specified by option_key is supported when saving a pixbuf using the module implementing self. See PixbufExt::save for more information about option keys.

Feature: v2_36

option_key

the name of an option

Returns

true if the specified option is supported

pub fn is_scalable(&self) -> bool

Returns whether this image format is scalable. If a file is in a scalable format, it is preferable to load it at the desired size, rather than loading it at the default size and scaling the resulting pixbuf to the desired size.

Returns

whether this image format is scalable.

pub fn is_writable(&self) -> bool

Returns whether pixbufs can be saved in the given format.

Returns

whether pixbufs can be saved in the given format.

pub fn set_disabled(&mut self, disabled: bool)

Disables or enables an image format. If a format is disabled, gdk-pixbuf won't use the image loader for this format to load images. Applications can use this to avoid using image loaders with an inappropriate license, see PixbufFormat::get_license.

disabled

true to disable the format self