Struct gdk_pixbuf::PixbufSimpleAnim

pub struct PixbufSimpleAnim(_, _);

An opaque struct representing a simple animation.

Implements

PixbufAnimationExt

Methods

impl PixbufSimpleAnim

fn new(width: i32, height: i32, rate: f32) -> PixbufSimpleAnim

Creates a new, empty animation.

width

the width of the animation

height

the height of the animation

rate

the speed of the animation, in frames per second

Returns

a newly allocated PixbufSimpleAnim

fn add_frame(&self, pixbuf: &Pixbuf)

Adds a new frame to self. The pixbuf must have the dimensions specified when the animation was constructed.

pixbuf

the pixbuf to add

fn set_loop(&self, loop_: bool)

Sets whether self should loop indefinitely when it reaches the end.

loop_

whether to loop the animation

fn get_loop(&self) -> bool

Gets whether self should loop indefinitely when it reaches the end.

Returns

true if the animation loops forever, false otherwise