Crate gdk_pixbuf

Modules

prelude
Traits inteded for blanket imports.

Structs

Continue
Continue calling the closure in the future iterations or drop it.

Error
A generic error capable of representing various error domains (types).

Object
The base class in the object hierarchy.

Pixbuf
This is the main structure in the gdk-pixbuf library. It is used to represent images. It contains information about the image's pixel data, its color space, bits per sample, width and height, and the rowstride (the number of bytes between the start of one row and the start of the next).

PixbufAnimation
An opaque struct representing an animation.

PixbufAnimationIter
An opaque struct representing an iterator which points to a certain position in an animation.

PixbufFormat

PixbufLoader
The PixbufLoader struct contains only private fields.

PixbufSimpleAnim
An opaque struct representing a simple animation.

TypedValue
A statically typed Value.

Value
A generic value capable of carrying various types.

Enums

Colorspace

InterpType

PixbufAlphaMode

PixbufError

PixbufRotation

Type
A GLib or GLib-based library type

Traits

Cast
Upcasting and downcasting support.

IsA
Declares the "is a" relationship.

PixbufAnimationExt
Trait containing all PixbufAnimation methods.

PixbufExt
Trait containing all Pixbuf methods.

PixbufLoaderExt
Trait containing all PixbufLoader methods.

PixbufSimpleAnimExt
Trait containing all PixbufSimpleAnim methods.

StaticType
Types that are supported by GLib dynamic typing.

ToValue
Converts to Value.