Crate gdk_pixbuf

Rust GDK-PixBuf bindings

Rust bindings and wrappers for GDK-Pixbuf, part of gtk-rs-core.

GDK-PixBuf 2.32 is the lowest supported version for the underlying library.

Minimum supported Rust version

Currently, the minimum supported Rust version is 1.56.0.

Documentation

Rust API - Stable

Rust API - Development

C API

GTK Installation instructions

Using

We recommend using crates from crates.io, as demonstrated here.

If you want to track the bleeding edge, use the git dependency instead:

[dependencies]
gdk-pixbuf = { git = "https://github.com/gtk-rs/gtk-rs-core.git", package = "gdk-pixbuf" }

Avoid mixing versioned and git crates like this:

# This will not compile
[dependencies]
gdk-pixbuf = "0.13"
gdk-pixbuf = { git = "https://github.com/gtk-rs/gtk-rs-core.git", package = "gdk-pixbuf" }

See Also

glib

gio

License

gdk-pixbuf is available under the MIT License, please refer to it.

Re-exports

pub use ffi;

pub use gio;

pub use glib;

Modules

prelude

Traits intended for blanket imports.

Structs

Pixbuf

A pixel buffer.

PixbufAnimation

An opaque object representing an animation.

PixbufAnimationIter

An opaque object representing an iterator which points to a certain position in an animation.

PixbufFormat

A PixbufFormat contains information about the image format accepted by a module.

PixbufFormatFlags

Flags which allow a module to specify further details about the supported operations.

PixbufLoader

Incremental image loader.

PixbufNonAnim

Implements

PixbufSimpleAnim

An opaque struct representing a simple animation.

Enums

Colorspace

This enumeration defines the color spaces that are supported by the gdk-pixbuf library.

InterpType

Interpolation modes for scaling functions.

PixbufAlphaMode

Control the alpha channel for drawables.

PixbufError

An error code in the GDK_PIXBUF_ERROR domain.

PixbufRotation

The possible rotations which can be passed to gdk_pixbuf_rotate_simple().