Crate gdk_pixbuf

Rust GDK-PixBuf bindings

Rust bindings and wrappers for GDK-Pixbuf, part of gtk-rs-core.

GDK-PixBuf 2.32 is the lowest supported version for the underlying library.

Minimum supported Rust version

Currently, the minimum supported Rust version is 1.56.0.

Documentation

Rust API - Stable

Rust API - Development

C API

GTK Installation instructions

Using

We recommend using crates from crates.io, as demonstrated here.

If you want to track the bleeding edge, use the git dependency instead:

[dependencies]
gdk-pixbuf = { git = "https://github.com/gtk-rs/gtk-rs-core.git", package = "gdk-pixbuf" }

Avoid mixing versioned and git crates like this:

[dependencies]
gdk-pixbuf = "0.13"
gdk-pixbuf = { git = "https://github.com/gtk-rs/gtk-rs-core.git", package = "gdk-pixbuf" }

See Also

glib

gio

License

gdk-pixbuf is available under the MIT License, please refer to it.

Re-exports

pub use ffi;

pub use gio;

pub use glib;

Modules

prelude

Traits intended for blanket imports.

Structs

Pixbuf

PixbufAnimation

PixbufAnimationIter

PixbufFormat

PixbufFormatFlags

PixbufLoader

PixbufNonAnim

PixbufSimpleAnim

Enums

Colorspace

InterpType

PixbufAlphaMode

PixbufError

PixbufRotation