Trait gdk_pixbuf::prelude::PixbufAnimationExt

pub trait PixbufAnimationExt: 'static {
    fn height(&self) -> i32;

    fn static_image(&self) -> Option<Pixbuf>;

    fn width(&self) -> i32;

    fn is_static_image(&self) -> bool;
}

Trait containing all PixbufAnimation methods.

Implementors

PixbufAnimation, PixbufSimpleAnim

Required methods

fn height(&self) -> i32

Queries the height of the bounding box of a pixbuf animation.

Returns

Height of the bounding box of the animation.

fn static_image(&self) -> Option<Pixbuf>

If an animation is really just a plain image (has only one frame), this function returns that image. If the animation is an animation, this function returns a reasonable thing to display as a static unanimated image, which might be the first frame, or something more sophisticated. If an animation hasn’t loaded any frames yet, this function will return None.

Returns

unanimated image representing the animation

fn width(&self) -> i32

Queries the width of the bounding box of a pixbuf animation.

Returns

Width of the bounding box of the animation.

fn is_static_image(&self) -> bool

If you load a file with PixbufAnimation::from_file() and it turns out to be a plain, unanimated image, then this function will return true. Use static_image() to retrieve the image.

Returns

true if the “animation” was really just an image