Struct gdk_pixbuf::PixbufFormat

pub struct PixbufFormat(_);

Methods

impl PixbufFormat

fn get_name(&self) -> Option<String>

Returns the name of the format.

Returns

the name of the format.

fn get_description(&self) -> Option<String>

Returns a description of the format.

Returns

a description of the format.

fn get_mime_types(&self) -> Vec<String>

Returns the mime types supported by the format.

Returns

a None-terminated array of mime types which must be freed with g_strfreev when it is no longer needed.

fn get_extensions(&self) -> Vec<String>

Returns the filename extensions typically used for files in the given format.

Returns

a None-terminated array of filename extensions which must be freed with g_strfreev when it is no longer needed.

fn is_writable(&self) -> bool

Returns whether pixbufs can be saved in the given format.

Returns

whether pixbufs can be saved in the given format.

fn is_scalable(&self) -> bool

Returns whether this image format is scalable. If a file is in a scalable format, it is preferable to load it at the desired size, rather than loading it at the default size and scaling the resulting pixbuf to the desired size.

Returns

whether this image format is scalable.

fn is_disabled(&self) -> bool

Returns whether this image format is disabled. See PixbufFormat::set_disabled.

Returns

whether this image format is disabled.

fn set_disabled(&self, disabled: bool)

Disables or enables an image format. If a format is disabled, gdk-pixbuf won't use the image loader for this format to load images. Applications can use this to avoid using image loaders with an inappropriate license, see PixbufFormat::get_license.

disabled

true to disable the format self

fn get_license(&self) -> Option<String>

Returns information about the license of the image loader for the format. The returned string should be a shorthand for a wellknown license, e.g. "LGPL", "GPL", "QPL", "GPL/QPL", or "other" to indicate some other license. This string should be freed with g_free when it's no longer needed.

Returns

a string describing the license of self.