gdk_pixbuf::subclass

Module prelude

Re-exports

pub use super::pixbuf_animation::PixbufAnimationImpl;

pub use super::pixbuf_animation::PixbufAnimationImplExt;

pub use super::pixbuf_animation_iter::PixbufAnimationIterImpl;

pub use super::pixbuf_animation_iter::PixbufAnimationIterImplExt;

pub use super::pixbuf_loader::PixbufLoaderImpl;

pub use super::pixbuf_loader::PixbufLoaderImplExt;

Traits

ActionGroupImpl

ActionGroupImplExt

ActionMapImpl

ActionMapImplExt

ApplicationImpl

ApplicationImplExt

AsyncInitableImpl

AsyncInitableImplExt

BoxedType

Trait for defining boxed types.

ClassStruct

Trait implemented by structs that implement a GObject C class struct.

DBusInterfaceImpl

DBusInterfaceImplExt

DBusProxyImpl

DBusProxyImplExt

DerivedObjectProperties

Trait containing only the property related functions of ObjectImpl. Implemented by the Properties macro. When implementing ObjectImpl you may want to delegate the function calls to this trait.

FileEnumeratorImpl

FileEnumeratorImplExt

FileImpl

FileImplExt

FileMonitorImpl

FileMonitorImplExt

IOStreamImpl

IOStreamImplExt

InitableImpl

InitableImplExt

InputStreamImpl

InputStreamImplExt

InstanceStruct

Trait implemented by structs that implement a GObject C instance struct.

InstanceStructExt

Trait implemented by any type implementing InstanceStruct to return the implementation, private Rust struct.

InterfaceStruct

Trait implemented by structs that implement a GTypeInterface C class struct.

IsImplementable

Trait for implementable interfaces.

IsSubclassable

Trait for subclassable class structs.

IsSubclassableExt

ListModelImpl

ListModelImplExt

ObjectClassSubclassExt

Extension trait for glib::Object’s class struct.

ObjectImpl

Trait for implementors of glib::Object subclasses.

ObjectImplExt

ObjectInterface

The central trait for defining a GObject interface.

ObjectInterfaceExt

ObjectInterfaceType

Type methods required for an ObjectInterface implementation.

ObjectSubclass

The central trait for subclassing a GObject type.

ObjectSubclassExt

Extension methods for all ObjectSubclass impls.

ObjectSubclassIsExt

Trait implemented by any type implementing ObjectSubclassIs to return the implementation, private Rust struct.

ObjectSubclassType

Type methods required for an ObjectSubclass implementation.

OutputStreamImpl

OutputStreamImplExt

RefCounted

SeekableImpl

SeekableImplExt

SharedType

Trait for defining shared types.

SocketControlMessageImpl

SocketControlMessageImplExt

TypeModuleImpl

TypeModuleImplExt

TypePluginImpl

TypePluginImplExt

TypePluginRegisterImpl

VfsImpl

VfsImplExt