gdk_pixbuf::subclass::prelude

Trait IsSubclassable

pub unsafe trait IsSubclassable<T>: IsSubclassableDefault<T>
where
    T: ObjectSubclass,
{
    // Provided methods
    fn class_init(class: &mut Class<Self>) { ... }
    fn instance_init(instance: &mut InitializingObject<T>) { ... }
}

Trait for subclassable class structs.

Provided Methods

fn class_init(class: &mut Class<Self>)

Override the virtual methods of this class for the given subclass and do other class initialization.

This is automatically called during type initialization and must call class_init() of the parent class.

fn instance_init(instance: &mut InitializingObject<T>)

Instance specific initialization.

This is automatically called during instance initialization and must call instance_init() of the parent class.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.