Trait gdk_pixbuf::PixbufAnimationExt

pub trait PixbufAnimationExt {
    fn get_width(&self) -> i32;

    fn get_height(&self) -> i32;

    fn get_iter(&self, start_time: TimeVal) -> PixbufAnimationIter;

    fn is_static_image(&self) -> bool;

    fn get_static_image(&self) -> Option<Pixbuf>;
}

Trait containing all PixbufAnimation methods.

Implementors

PixbufAnimation, PixbufSimpleAnim

Required methods

fn get_width(&self) -> i32

Queries the width of the bounding box of a pixbuf animation.

Returns

Width of the bounding box of the animation.

fn get_height(&self) -> i32

Queries the height of the bounding box of a pixbuf animation.

Returns

Height of the bounding box of the animation.

fn get_iter(&self, start_time: TimeVal) -> PixbufAnimationIter

Get an iterator for displaying an animation. The iterator provides the frames that should be displayed at a given time. It should be freed after use with gobject::Object::unref.

start_time would normally come from g_get_current_time, and marks the beginning of animation playback. After creating an iterator, you should immediately display the pixbuf returned by PixbufAnimationIterExt::get_pixbuf. Then, you should install a timeout (with g_timeout_add) or by some other mechanism ensure that you'll update the image after PixbufAnimationIterExt::get_delay_time milliseconds. Each time the image is updated, you should reinstall the timeout with the new, possibly-changed delay time.

As a shortcut, if start_time is None, the result of g_get_current_time will be used automatically.

To update the image (i.e. possibly change the result of PixbufAnimationIterExt::get_pixbuf to a new frame of the animation), call PixbufAnimationIterExt::advance.

If you're using PixbufLoader, in addition to updating the image after the delay time, you should also update it whenever you receive the area_updated signal and PixbufAnimationIterExt::on_currently_loading_frame returns true. In this case, the frame currently being fed into the loader has received new data, so needs to be refreshed. The delay time for a frame may also be modified after an area_updated signal, for example if the delay time for a frame is encoded in the data after the frame itself. So your timeout should be reinstalled after any area_updated signal.

A delay time of -1 is possible, indicating "infinite."

start_time

time when the animation starts playing

Returns

an iterator to move over the animation

fn is_static_image(&self) -> bool

If you load a file with PixbufAnimation::new_from_file and it turns out to be a plain, unanimated image, then this function will return true. Use PixbufAnimationExt::get_static_image to retrieve the image.

Returns

true if the "animation" was really just an image

fn get_static_image(&self) -> Option<Pixbuf>

If an animation is really just a plain image (has only one frame), this function returns that image. If the animation is an animation, this function returns a reasonable thing to display as a static unanimated image, which might be the first frame, or something more sophisticated. If an animation hasn't loaded any frames yet, this function will return None.

Returns

unanimated image representing the animation

Loading content...